+++ perf-event/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.1] - 2023-05-19
 ### Added
 - Expose the `IOC_SET_BPF` ioctl as `Counter::set_bpf`.
 - Add `Event::update_attrs_with_data` to allow events to store references to